% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oarg_decode.R
\name{oarg_decode}
\alias{oarg_decode}
\title{Decode and store OpenAIRE Research Graph files}
\usage{
oarg_decode(oaire = NULL, limit = NULL, records_path = NULL,
  verbose = TRUE)
}
\arguments{
\item{oaire}{compressed json file}

\item{limit}{number of records to be decoded}

\item{records_path}{directory for the xml files}

\item{verbose}{print some information on what is going on}
}
\value{
files
}
\description{
Decode and store OpenAIRE Research Graph files
}
