% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oarg_linked_ftxts.R
\name{oarg_linked_ftxt}
\alias{oarg_linked_ftxt}
\title{Get full-text links}
\usage{
oarg_linked_ftxt(doc)
}
\arguments{
\item{doc}{document of class `xml_document` from the OpenAIRE research graph publication subset}
}
\value{
The result is a tibble. Here are the returned columns and descriptions:

\tabular{ll}{
\code{access_right}    \tab Availability of the resource \cr
\code{collected_from}  \tab Name of the metadata source \cr
\code{instance_type}   \tab Type of linked resource \cr
\code{hosted_by}       \tab Name of the Repository or Journal \cr
\code{web_urls}        \tab List-columns with links to the resource
}
}
\description{
OpenAIRE keep tracks of full-text links to publications including access level.
This parser retrieves links and corresponding metadata about the host of a resource.
}
\examples{
\dontrun{
# sample file delivered with this package
dump_eg <- system.file("extdata", "multiple_projects.xml", package = "openairegraph")
# load xml file
my_record <- xml2::read_xml(dump_eg)
# parse full text links
oarg_linked_ftxt(my_record)
}

}
