% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oarg_publications_md.R
\name{oarg_publications_md}
\alias{oarg_publications_md}
\title{Get publication metadata}
\usage{
oarg_publications_md(doc)
}
\arguments{
\item{doc}{document of class `xml_document` from the OpenAIRE research graph publication subset}
}
\value{
data.frame with publication metadata
}
\description{
Get publication metadata
}
