% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oarg_linked_affiliations.R
\name{oarg_linked_affiliations}
\alias{oarg_linked_affiliations}
\title{Parse affiliations}
\usage{
oarg_linked_affiliations(doc)
}
\arguments{
\item{doc}{document of class `xml_document` from the OpenAIRE research graph publication subset}
}
\value{
data.frame with affiliations linked to a publication
}
\description{
Obtain affiliation metadata from OpenAIRE publications
}
