% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oarg_linked_ftxts.R
\name{oarg_linked_ftxt}
\alias{oarg_linked_ftxt}
\title{Get full-text links}
\usage{
oarg_linked_ftxt(doc)
}
\arguments{
\item{doc}{document of class `xml_document` from the OpenAIRE research graph publication subset}
}
\value{
data.frame with full-text links to a publication inlcuding access level
}
\description{
Obtain full-text links to OpenAIRE publications including access level
}
