% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oarg_linked_projects.R
\name{oarg_linked_projects}
\alias{oarg_linked_projects}
\title{Parse projects}
\usage{
oarg_linked_projects(doc)
}
\arguments{
\item{doc}{document of class `xml_document` from the OpenAIRE research graph publication subset}
}
\value{
data.frame with project information linked to a publication
}
\description{
Obtain project metadata linked to OpenAIRE publications
}
